grpc/grpc-metadata/client-py/requirements.txt
grpcio>=1.36.1
grpcio-status>=1.36.1
grpcio-tools>=1.36.1
googleapis-common-protos>=1.53.0

